backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pandas
scikit-learn
joblib
gunicorn
pytesseract
Pillow
pdf2image
reportlab
